Error in API fixed, but still no cigar
Fixes the following reported error in "Error List" tab:

"The parameters to the Event method do not match the parameters to the
WriteEvent method. This may cause the event to be displayed incorrectly"

Still no cigar regarding the HeartBeat issue in SFX shown here:
https://ibin.co/3CFCBVCOWHya.png

Note: Git diff shows annoying ^M diffs. These seems to be false as they are
not possible to remove in Vim (i.e. it's hence believed git diff is the
cause of this).