Part #42
OnOpenAsync re-written verbatim from lab. Does the following:

* Calls the newly added VotingService::LoadConfiguration() method
* Instantiates a new _heathTimer which replaces the one instantiated in the
  constructor

Side note:
==========
During implementing this part it was noticed that "Diagnostic Event" was
picking up the heart-beat events. Fairly slowly, hence possibly explaining
why this was this was possibly not noticed before.

The view doesn't correspond to the documentation however and the events
(aggregated as status) is instead seen, but in a specific partition:
https://imagebin.ca/v/3CFzDsiVk5wf